platform: add windows platform support
1. add windows platform support
2. rename pal.h -> PAL.h
3. add gamepad: todo

+++ src/KeyHistory.cpp
@@ -46,9 +46,9 @@ void KeyHistory::__gameKeyVisualImpl() {
 
     style.ItemSpacing = ImVec2(spacing, spacing);
 
-    int areaHeight = ImGui::GetWindowHeight() / 3;
-    int areaWidth = ImGui::GetWindowWidth() / 3;
-    int bWidth = areaWidth / 4, bHeight = bWidth;
+    float areaHeight = ImGui::GetWindowHeight() / 3;
+    float areaWidth = ImGui::GetWindowWidth() / 3;
+    float bWidth = areaWidth / 4, bHeight = bWidth;
 
     {
         int firstLineAreaStart = areaWidth / 2 - bWidth / 2;
